requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
psycopg2-binary>=2.9,<3
requests>=2.32,<3

